Unit tests PlayerUtil
