+++ src/main/java/com/polifono/util/PasswordUtil.java
@@ -1,0 +1,15 @@
+package com.polifono.util;
+
+import java.security.SecureRandom;
+
+import javax.annotation.Nonnull;
+
+import org.springframework.security.crypto.bcrypt.BCrypt;
+
+public class PasswordUtil {
+
+    public static String encryptPassword(@Nonnull final String rawPassword) {
+        final String salt = BCrypt.gensalt(10, new SecureRandom());
+        return BCrypt.hashpw(rawPassword, salt);
+    }
+}
